refactor: servicio de autenticación
