+++ src/app/app.routes.ts
@@ -1,5 +1,5 @@
 import { Routes } from '@angular/router';
-import { LoginComponent } from './modules/layout/login/component/login.component';
+import { LoginComponent } from './modules/layout/login/login.component';
 import { RegisterComponent } from './modules/layout/register/register.component';
 import { TemplateComponent } from './modules/layout/template/template.component';
 import { ProfileComponent } from './modules/layout/profile/profile.component';
